full single-snake
