+++ Networks_Lab4(Snake)/src/main/java/controllers/GameBoardController.java
@@ -1,21 +1,19 @@
 package controllers;
 
-import exceptions.NoEmptyCellException;
 import gui.GameBoardView;
 import models.GameBoardModel;
 import models.GameProcess;
+import models.ServerGameProcess;
 import observers.Observer;
 import protocols.SnakeProto.*;
 
 import javax.swing.*;
-import java.util.Objects;
-import java.util.Random;
 
 public class GameBoardController implements Observer {
     private GameBoardView gameBoardView;
     private GameBoardModel gameBoardModel;
     private StatusController statusController;
-    private Random random = new Random();
+    private GameProcess gameProcess;
 
     public GameBoardController(GameBoardView gameBoardView, GameBoardModel gameBoardModel, StatusController statusController){
         this.gameBoardView = gameBoardView;
@@ -28,50 +26,9 @@ public class GameBoardController implements Observer {
     }
 
     public void createGame(){
-        gameBoardModel.createBoard(50, 50);
-        gameBoardView.setGameBoardSize(50, 50);
-        gameBoardView.setCells(gameBoardModel.getBoardCells());
-        GameState.Snake snake = null;
-//        try {
-//            snake = createSnake();
-//            gameBoardModel.addSnake(snake);
-//            gameBoardModel.changeSnakeDirection(snake.getPlayerId(), snake.getHeadDirection());
-//        } catch (NoEmptyCellException e) {
-//            e.printStackTrace();
-//        }
-        snake = GameState.Snake.newBuilder()
-                .setPlayerId(0)
-                .setHeadDirection(Direction.UP)
-                .setState(GameState.Snake.SnakeState.ALIVE)
-                .addPoints(coord(5, 5))
-                .addPoints(coord(3,0))
-                .addPoints(coord(0, 3))
-                .addPoints(coord(5, 0))
-                .build();
-        gameBoardModel.addSnake(snake);
-        gameBoardModel.changeSnakeDirection(snake.getPlayerId(), snake.getHeadDirection());
-        GameProcess gameProcess = new GameProcess(gameBoardModel, gameBoardView, 1000);
-        Thread gameProcessThread = new Thread(gameProcess);
-        gameProcessThread.start();
-    }
-
-    public GameState.Snake createSnake() throws NoEmptyCellException {
-        GameState.Coord snakeHead = gameBoardModel.getFreeCell();
-        Direction headDirection = Direction.forNumber(random.nextInt(4) + 1);
-        GameState.Coord snakeBody = null;
-        switch (headDirection){
-            case UP -> snakeBody = coord(snakeHead.getX(), snakeHead.getY() + 1);
-            case DOWN -> snakeBody = coord(snakeHead.getX(), snakeHead.getY() - 1);
-            case LEFT -> snakeBody = coord(snakeHead.getX() + 1, snakeHead.getY());
-            case RIGHT -> snakeBody = coord(snakeHead.getX() - 1, snakeHead.getY());
-        }
-        return GameState.Snake.newBuilder()
-                .setPlayerId(0)
-                .setHeadDirection(headDirection)
-                .setState(GameState.Snake.SnakeState.ALIVE)
-                .addPoints(snakeHead)
-                .addPoints(snakeBody)
-                .build();
+        gameProcess = new ServerGameProcess(gameBoardModel, gameBoardView, 200);
+        gameProcess.createGame();
+        gameProcess.start();
     }
 
     private GameState.Coord coord(int x, int y) {
